Moved AlwaysAvailableAvailabilityStrategy to availability strategy
