+++ airbyte-cdk/python/airbyte_cdk/sources/streams/concurrent/availability_strategy.py
@@ -65,3 +65,22 @@ class AbstractAvailabilityStrategy(ABC):
         :param logger: logger object to use
         :return: A StreamAvailability object describing the stream's availability
         """
+
+
+class AlwaysAvailableAvailabilityStrategy(AbstractAvailabilityStrategy):
+    """
+    An availability strategy that always indicates a stream is available.
+
+    This strategy is used to avoid breaking changes and serves as a soft
+    deprecation of the availability strategy, allowing a smoother transition
+    without disrupting existing functionality.
+    """
+
+    def check_availability(self, logger: logging.Logger) -> StreamAvailability:
+        """
+        Checks stream availability.
+
+        :param logger: logger object to use
+        :return: A StreamAvailability object describing the stream's availability
+        """
+        return StreamAvailable()
